refactor: 피드백 반영1
