+++ src/main/java/nextstep/subway/application/LineService.java
@@ -1,11 +1,14 @@
 package nextstep.subway.application;
 
+import nextstep.subway.domain.Distance;
 import nextstep.subway.domain.Line;
 import nextstep.subway.domain.LineRepository;
+import nextstep.subway.domain.Section;
 import nextstep.subway.domain.Station;
 import nextstep.subway.dto.LineRequest;
 import nextstep.subway.dto.LineResponse;
 import nextstep.subway.dto.LineUpdateRequest;
+import nextstep.subway.dto.SectionRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,7 +19,7 @@ import java.util.stream.Collectors;
 @Service
 public class LineService {
 
-    private final String NOT_FOUND_BY_ID ="ID로 찾을 수 없습니다.";
+    private final String NOT_FOUND_BY_ID = "ID로 찾을 수 없습니다.";
     private final LineRepository lineRepository;
     private final StationService stationService;
 
@@ -54,12 +57,21 @@ public class LineService {
     @Transactional
     public void updateLine(Long id, LineUpdateRequest lineUpdateRequest) {
         Line line = lineRepository.findById(id)
-            .orElseThrow(() -> new IllegalArgumentException(NOT_FOUND_BY_ID));
+            .orElseThrow(() -> new IllegalArgumentException(NOT_FOUND_BY_ID + "id: " + id));
         line.updateLine(lineUpdateRequest.getName(), lineUpdateRequest.getColor());
     }
 
     @Transactional
     public void deleteLineById(Long id) {
         lineRepository.deleteById(id);
+    }
+
+    @Transactional
+    public LineResponse addSection(Long id, SectionRequest sectionRequest) {
+        Station upStation = stationService.findById(sectionRequest.getUpStationId());
+        Station downStation = stationService.findById(sectionRequest.getDownStationId());
+        Line line = lineRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+        line.addSection(Section.of(upStation, downStation, Distance.from(sectionRequest.getDistance())));
+        return LineResponse.of(line);
     }
 }
