+++ src/main/java/nextstep/subway/domain/Section.java
@@ -57,7 +57,7 @@ public class Section {
         return new Section(upStation, downStation, distance);
     }
 
-    public void toLine(Line line) {
+    public void addLine(Line line) {
         this.line = line;
     }
